Add config to allow this repository to be used as a GitHub Action
This configuration exposes a composite action that will clone Artichoke
and run the generate_third_party gem against its lockfile.

The action takes a git ref for the artichoke/artichoke repository and a
target triple to generate the license listing for.

The action sets the generated license listing as an output.