Add Ruby script for downloading cargo-about
Downloading release information with curl and piping it to jq is flaky
in CI and in the nightly builder, sometimes resulting in 403 errors that
curl treats as unrecoverable.

curl has a `--retry-all-errors` flag but it is unavailable on Windows or
Ubuntu 20.04.

Add a small Ruby script using only the standard library that is
installed as part of the `generate_third_party` gem. This script
features retries on network errors, logging with log GitHub Actions log
groups, and automatically adds the resulting cargo-about binaries to
$PATH.

The script takes a platform as its first argument which conveniently
matches the `runner.os` context variable in a GitHub Actions context.